Disable Windows and MacOS
Windows and MacOS templates need KVM to work, the problem is that KVM is only available on Linux. The second problem is that KVM needs to be installed on the host server, but most Coolify users use a VPS to host it, and almost all VPS providers block KVM access or installation on the host because the hole server is already KVM virtualized.